requirements.txt
openai
PyMuPDF
requests
faiss-cpu
numpy
python-dotenv